Check if the user has integer a double or integer value

+++ app/src/main/java/com/example/pottery/viewModels/FormulaViewModel.kt
@@ -75,5 +75,5 @@ class FormulaViewModel(application: Application) : AndroidViewModel(application)
     fun updateItem(item: Item) = repository.updateItem(item)
 
 }
-data class NewItem(var code :String,var material:String, var amount : Int)
+data class NewItem(var code :String,var material:String, var amount : Double)
